+++ src/hera/resources.py
@@ -4,7 +4,7 @@ from typing import Optional, Union
 from pydantic import BaseModel, root_validator, validator
 
 from hera.validators import validate_storage_units
-from hera.volumes import EmptyDirVolume, ExistingVolume, Volume
+from hera.volumes import EmptyDirVolume, ExistingVolume, SecretVolume, Volume
 
 
 class Resources(BaseModel):
@@ -41,6 +41,7 @@ class Resources(BaseModel):
     volume: Optional[Volume] = None
     existing_volume: Optional[ExistingVolume] = None
     empty_dir_volume: Optional[EmptyDirVolume] = None
+    secret_volume: Optional[SecretVolume] = None
 
     overwrite_maxs: bool = True